refactor: Update viewModel DSL import
This commit updates the viewModel DSL import for Koin to use `koin.core.module.dsl.viewModel` instead of the deprecated `koin.compose.viewmodel.dsl.viewModel` and its variants.

This change ensures compatibility with the latest Koin API and streamlines the viewModel declaration in different feature modules. Specifically, the import is updated in:

- `features/setting/src/commonMain/kotlin/com/mshdabiola/setting/SettingModule.kt`
- `features/detail/src/commonMain/kotlin/com/mshdabiola/detail/DetailModule.kt`
- `features/main/src/commonMain/kotlin/com/mshdabiola/main/MainModule.kt`
- `app/src/commonMain/kotlin/com/mshdabiola/skeletonapp/di/ApplicationModule.kt`