chore: Improve GenerateBaselineProfile
The `GenerateBaselineProfile` benchmark has been updated to enhance its simulation of real-world usage.

The key change is to repeat the flow of going to the detail screen, adding a note, and going back two times instead of once. This ensures the most important parts of the app are captured within the profile.