chore: Update ProfileCardPreview and Main
This commit removes the `expect` and `actual` keywords from the `ProfileCardPreview` composable function,
making it platform-agnostic.

Additionally, it initializes Koin in the `main` function for the Wasm target and includes Kermit logger initialization.
This allows dependency injection and logging in the Wasm environment.

The changes also include removing platform-specific `ProfileCardPreview` implementations.
Instead, the `ProfileCardPreview` function is now defined once and used for all platforms.