feat: Add Splash screen to WebAssembly build
This commit adds a splash screen to the WebAssembly build of the application.

The splash screen is displayed for 2 seconds before the main application is loaded. This provides a more polished user experience and allows time for the application to initialize.

The splash screen is implemented using a simple `SplashScreen` composable function that displays a logo and some text. It is displayed in a `Box` composable function that also contains the main application. The `show` state variable is used to control the visibility of the splash screen. When the `show` state variable is `true`, the splash screen is displayed. When the `show` state variable is `false`, the main application is displayed.

The `LaunchedEffect` composable function is used to delay the loading of the main application by 2 seconds. This ensures that the splash screen is displayed for the desired amount of time.

The `mainApp` function has been created to encapsulate the splash screen and main application logic. This makes it easier to manage the splash screen and main application.