refactor: Remove lint module and dependency
This commit removes the `lint` module and its dependency from the build configuration.

The `lint` module was previously included in `settings.gradle.kts` and used as a `lintPublish` dependency in the `designsystem` module.

With this change, the `lint` module and its associated configurations are no longer part of the project. Lint checks will now rely on the default Android linting configurations or other configured linting tools like Ktlint.