refactor: Organize imports in ApplicationModule and MainAppViewModel
This commit rearranges the import statements in `ApplicationModule.kt` and `MainAppViewModel.kt` to group related imports and adhere to a standard format, improving code readability and consistency.

Additionally, the import of `HydraulicAppViewModel` in `ApplicationModule.kt` was moved before the `detailModule` import.

In `HydraulicApp.kt`, the import for `HydraulicAppViewModel` was also moved. The import for `Icons.Default.ArrowBackIosNew` was removed and replaced with the custom `HyaIcons.ArrowBack` in the `HyaTopAppBar` for the setting screen.