build: update app version to 1.2.7
This commit updates the application's version to 1.2.7. This includes updating the `versionCode` to 127 and the `versionName` to 1.2.7 in the `libs.versions.toml` file.

Additionally, the displayed version in the JVM desktop application's window title and the version information in the `fossReliantRelease-badging.txt` file have been updated to reflect the new version.