refactor: Update splash screen and string resources
This commit refactors the `SplashScreen` composable to accept the application name as a parameter, defaulting to "HydraulicApp". This makes the component more reusable and less dependent on specific string resources.

Additionally, the `appName` and `type` string resource properties in the `designsystem` module have been renamed to `appName2` and `type2` respectively.