Rename `make site` to `make copied-site` in Hydraulic Conveyor workflow
This commit renames the `make site` command to `make copied-site`
in the Hydraulic Conveyor pre-release GitHub Actions workflow.
It also removes the deployment steps for GitHub Pages, as these are
now handled by a separate workflow.

Additionally, the commit updates the `ci.conveyor.conf` file to
include the number of commits in the current branch in the app
configuration.