Refactor: Rename CollectSarifPlugin to CiTaskPlugin and add BumpConveyorRevisionTask
This commit refactors the `CollectSarifPlugin` to `CiTaskPlugin`.
The Sarif merging tasks for lint and detekt are commented out in the new plugin.
A new task, `bumpConveyorRevision`, of type `BumpConveyorRevisionTask` is registered within the `CiTaskPlugin`. This task is responsible for reading, incrementing, and updating the Conveyor revision number in the `.github/workflows/.revision-version` and `ci.conveyor.conf` files.