ci: Add setVersionFromTag and updateFossReliantReleaseBadging to HydraulicConveyor workflow
This commit updates the `HydraulicConveyor.yml` GitHub Actions workflow.

It adds a step to run the `setVersionFromTag` Gradle task, using the `github.ref_name` as the `newVersionName`.

It also adds a step to run the `./gradlew updateFossReliantReleaseBadging` command.

Finally, it adds a step to create a pull request titled "chore: Bump Version name and code" using `peter-evans/create-pull-request`. This PR will be created against the `develop` branch.