feat: Add SetVersionTagTask to update version info
This commit introduces a new Gradle task `SetVersionTagTask`.

This task updates version information in `BuildConfig.kt` and `gradle/libs.versions.toml`.
It sets `VERSION_NAME` and `VERSION_CODE_DESKTOP` in `BuildConfig.kt` and `versionName` and `desktopCode` in `libs.versions.toml` to a new input tag.

The task is registered in `CiTaskPlugin.kt` as `setVersionTag`.