Refactor: Rename scaffold test tags
This commit updates the test tags in `KmtScaffoldTestTags.kt` to use a more specific prefix.

The prefix for all test tags has been changed from a generic one (e.g., "scaffold:", "fab:", "drawer:") to "SynScaffold:". Additionally, the format for dynamic navigation item tags was updated to use a colon as a separator instead of an underscore.